Crypto library movement and changes to content helper interfaces
