+++ pkg/encryption/reader.go
