+++ pkg/encryption/utils/utils.go
